Dummy implementation of sgl-fft-viz.  Pausing to refactor
Refactoring as described in project.org to clean up how
buffers and textures  are initialized, filled, refilled, and cleaned
up.

Currently, refill is done by cleaning up and re-initializing, which was a
shortcut while I got that stuff working.  Now that it's working well, I need to
go back and fine tune it and make it more flexible.